% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradfps.R
\name{gradfps_prox_omd}
\alias{gradfps_prox_omd}
\title{The GradFPS algorithm for sparse principal component analysis}
\usage{
gradfps_prox_omd(
  S,
  d,
  lambda,
  delta,
  x0 = NULL,
  lr = 0.001,
  maxiter = 100L,
  control = list()
)
}
\arguments{
\item{S}{The sample covariance matrix.}

\item{d}{The number of sparse principal components to seek.}

\item{lambda}{Sparsity parameter.}

\item{delta}{Penalty parameter.}

\item{x0}{Initial value for the projection matrix. If \code{NULL}, a
default value based on the regular PCA will be used.}

\item{maxiter}{Maximum number of iterations.}

\item{control}{Additional parameters to control the optimization process.}
}
\description{
The GradFPS algorithm for sparse PCA using online mirror descent method.
It solves the optimization problem
\deqn{\min\quad-tr(SX) + \lambda ||X||_{1,1} + \frac{\delta}{2}||X||_{r,r}^2,}{min  -tr(AX) + \lambda * ||X||_{1,1} + (\delta / 2) * ||X||_{r,r}^2,}
\deqn{s.t.\quad X\in \mathcal{F}^d,}{s.t. X in F^d,}
where \eqn{r = 1 + 1 / (\log(p) - 1)}{r = 1 + 1 / (log(p) - 1)}, and
\eqn{\mathcal{F}^d}{F^d} is the Fantope.
}
\details{
The \code{control} argument is a list that can supply any of the
following parameters:

\describe{
\item{\code{eps_abs}}{Absolute precision parameter. Default is \code{1e-3}.}
\item{\code{eps_rel}}{Relative precision parameter. Default is \code{1e-3}.}
\item{\code{eig_method}}{Method to compute eigenvalues. Possible values are
                         \code{"spectra"} and \code{"lapack"}. Default is
                         \code{"spectra"}.}
\item{\code{fan_maxinc}}{Maximum number of incremental eigenvalues to compute
                         in each iteration of the Fantope proximal operator.
                         Default is 100.}
\item{\code{fan_maxiter}}{Maximum number of iterations in the Fantope proximal operator.
                          Default is 10.}
\item{\code{verbose}}{Level of verbosity. Default is 1.}
}
}
