% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{view_matrix}
\alias{view_matrix}
\alias{view_evec}
\title{Visualization of Matrices}
\usage{
view_matrix(
  mat,
  xlab = "",
  ylab = "",
  legend_title = "Coefficient",
  bar_height = 10,
  font_size = 20,
  padding = 0
)

view_evec(
  evecs,
  xlab = "Index of Variables",
  ylab = "Index of PCs",
  legend_title = "Factor\\nLoading",
  asp = 0.2,
  bar_height = 6,
  font_size = 20
)
}
\arguments{
\item{mat, evec}{The matrix to be visualized.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{legend_title}{Title of the colorbar legend.}

\item{bar_height}{Height of the colorbar.}

\item{font_size}{Base font size for the plot.}

\item{padding}{The padding between the axis border and the plot area.}

\item{asp}{Aspect ratio of the plot.}
}
\description{
\code{view_matrix()} visualizes a general matrix by mapping its elements to
colors. \code{view_evec()} visualizes a set of eigenvectors.
}
\examples{
set.seed(123)
x = matrix(rnorm(200), 10, 20)
view_matrix(x)

sigma1 = matrix(0.8, 20, 20) + 0.2 * diag(20)
sigma2 = matrix(0.6, 50, 50) + 0.4 * diag(50)
s1 = stats::rWishart(1, 100, sigma1)[, , 1]
s2 = stats::rWishart(1, 100, sigma2)[, , 1]
s = as.matrix(Matrix::bdiag(s1, s2))
view_matrix(s)

v = eigen(s, symmetric = TRUE)$vectors[, 1:5]
view_evec(v)
}
\author{
Yixuan Qiu \url{https://statr.me}
}
